google_sheets_api/requirements.txt
Django==4.0.4
google-api-python-client==2.48.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
gspread==5.3.2
schedule==1.1.0
